cron/pyproject.toml
[project]
name = "agos-cron"
version = "0.1.0"
description = "Scraping pipeline for financial and macro data"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "curl-cffi>=0.7.1",
    "lxml>=5.1.0",
    "pandas>=2.0.0",
    "SQLAlchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
